requirements-minimal.txt
Flask>=3.0.0
Flask-CORS>=4.0.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
Werkzeug>=3.0.0
